get and post fix

+++ howMoney/Models/AlertDTO.cs
@@ -4,7 +4,7 @@
     {
         public string asset_name { get; set; } = string.Empty;
         public string currency { get; set; } = string.Empty;
-        public int value { get; set; } = 0;
+        public double value { get; set; } = 0.0;
 
     }
 }
