+++ howMoney/Models/AlertWithEmailDTO.cs
@@ -5,7 +5,7 @@
         public string email { get; set; } = string.Empty;
         public string asset_name { get; set; } = string.Empty;
         public string currency { get; set; } = string.Empty;
-        public int value { get; set; } = 0;
+        public double value { get; set; } = 0.0;
 
         public AlertWithEmailDTO()
         {
